@@ -33,12 +33,23 @@ import (
 	"testing"
 
 	"github.com/fairwindsops/pluto/v5/pkg/api"
+        "github.com/fairwindsops/pluto/v5/pkg/kube"
 	"github.com/stretchr/testify/assert"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
+	testclient "k8s.io/client-go/kubernetes/fake"
 	"k8s.io/client-go/rest"
 )
+
+var kubeClient *kube.Kube
+
+func getMockConfigInstance() *kube.Kube {
+        kubeClient = &kube.Kube{
+                Client: testclient.NewSimpleClientset(),
+        }
+        return kubeClient
+}
 
 var (
 	helmSecret = v1.Secret{
@@ -185,10 +196,10 @@ func newMockHelm(namespace string) *Helm {
 	}
 }
 
-func newBadKubeClient() (k *kube) {
+func newBadKubeClient() (k *kube.Kube) {
 	conf := new(rest.Config)
 	conf.Host = "127.0.0.1:9999"
-	k = new(kube)
+	k = new(kube.Kube)
 	k.Client, _ = kubernetes.NewForConfig(conf)
 	return
 }
