@@ -12,23 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package helm
+package kube
 
 import (
 	"os"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	testclient "k8s.io/client-go/kubernetes/fake"
 	_ "k8s.io/client-go/plugin/pkg/client/auth"
 )
-
-func getMockConfigInstance() *kube {
-	kubeClient = &kube{
-		Client: testclient.NewSimpleClientset(),
-	}
-	return kubeClient
-}
 
 func Test_getKubeClient(t *testing.T) {
 	tests := []struct {
@@ -59,7 +51,7 @@ func Test_getKubeClient(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			os.Setenv("KUBECONFIG", tt.kubeConfig)
-			_, err := getKubeClient(tt.kubeContext)
+                        _, err := GetConfig(tt.kubeContext)
 			if tt.wantErr {
 				assert.Error(t, err)
 			} else {
